Further work for reorganisation
- Mostly working now

+++ build.sh
@@ -2,9 +2,19 @@
 
 set -e
 debug=1
+dev=1
 
 mkdir -p build
-rm -rf build/*
+
+if [ $dev ]; then
+  echo "Linking in backend files for 'dev' deploy"
+  (cd build ; for f in ../app/backend/*.hs ; do ln -s $f .; done )
+  (cd build ; ln -s r-json.hs r-json.cgi)
+  mkdir -p build/tmp build/user-files build/cached
+else
+  rm -rf build/*
+fi
+
 
 cp -r app/html/ build
 cp -r app/images build
@@ -12,7 +22,6 @@ cp -r app/images build
 #echo "Combining css and minifying..."
 #cat css/bootstrap-tour.min.css css/dge.css css/venn.css css/slick.grid.css | cleancss > build/main.min.css
 cp -r app/css build
-
 
 echo "Compiling CoffeeScript and bundling all js..."
 for f in app/js/*-req.coffee; do
@@ -29,3 +38,9 @@ for f in app/js/*-req.coffee; do
       rm $b
   fi
 done
+
+
+
+if [ $dev ]; then
+  echo "Now run ./server.py"
+fi
